move LANGUAGE variable to config/app.js
