+++ src/config/app.js
@@ -1,3 +1,23 @@
+export const LANGUAGE = [
+    { code: 'en', name: 'English', en: 'English' },
+    { code: 'bn', name: 'Bengali', en: 'Bengali' },
+    { code: 'cs', name: 'Česky', en: 'Czech' },
+    { code: 'de', name: 'Deutsch', en: 'German' },
+    { code: 'el', name: 'Ελληνικά', en: 'Greek' },
+    { code: 'es', name: 'Español', en: 'Spanish' },
+    { code: 'fr', name: 'Français', en: 'French' },
+    { code: 'id', name: 'Bahasa Indonesia', en: 'Indonesian' },
+    { code: 'it', name: 'Italiano', en: 'Italian' },
+    { code: 'ja', name: '日本語', en: 'Japanese' },
+    { code: 'nl', name: 'Nederlands', en: 'Dutch' },
+    { code: 'pl', name: 'Polski', en: 'Polish' },
+    { code: 'pt', name: 'Português', en: 'Portuguese' },
+    { code: 'ru', name: 'Русский', en: 'Russian' },
+    { code: 'uk', name: 'Український', en: 'Ukrainian' },
+    { code: 'zh', name: '中文(简体)', en: 'Chinese Simplified' },
+    { code: 'zh_TW', name: '中文(台灣)', en: 'Chinese Traditional' },
+];
+
 export const FIAT_CURRENCIES = [
     'usd',
     'aed',
